requirements.txt
transformers
torch
huggingface_hub
edge-tts
requests
python-dotenv
feedparser
git+https://github.com/JustAnotherArchivist/snscrape.git
praw
pandas
pytrends
google-api-python-client
python-dotenv
